+++ integration/spark/src/main/common/java/io/openlineage/spark/agent/lifecycle/SparkSQLExecutionContext.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.SparkContext;
@@ -36,7 +37,6 @@ import org.apache.spark.scheduler.SparkListenerJobStart;
 import org.apache.spark.sql.SparkSession;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
 import org.apache.spark.sql.execution.QueryExecution;
-import org.apache.spark.sql.execution.SQLExecution;
 import org.apache.spark.sql.execution.SparkPlan;
 import org.apache.spark.sql.execution.WholeStageCodegenExec;
 import org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionEnd;
@@ -61,29 +61,64 @@ public class SparkSQLExecutionContext implements ExecutionContext {
   private final OpenLineage openLineage =
       new OpenLineage(OpenLineageClient.OPEN_LINEAGE_CLIENT_URI);
 
+  private AtomicBoolean started = new AtomicBoolean(false);
+  private AtomicBoolean finished = new AtomicBoolean(false);
+  private Optional<Integer> jobId = Optional.empty();
+
   public SparkSQLExecutionContext(
       long executionId,
       OpenLineageContext sparkContext,
+      QueryExecution queryExecution,
       List<QueryPlanVisitor<LogicalPlan, OpenLineage.OutputDataset>> outputDatasetSupplier,
       List<QueryPlanVisitor<LogicalPlan, OpenLineage.InputDataset>> inputDatasetSupplier) {
     this.executionId = executionId;
     this.sparkContext = sparkContext;
-    this.queryExecution = SQLExecution.getQueryExecution(executionId);
+    this.queryExecution = queryExecution;
     this.outputDatasetSupplier = outputDatasetSupplier;
     this.inputDatasetSupplier = inputDatasetSupplier;
     this.jobMetrics = JobMetricsHolder.getInstance();
   }
 
-  public void start(SparkListenerSQLExecutionStart startEvent) {}
+  public void start(SparkListenerSQLExecutionStart startEvent) {
+    log.debug("SparkListenerSQLExecutionStart - executionId: {}", startEvent.executionId());
+    startEvent(startEvent.time());
+  }
 
-  public void end(SparkListenerSQLExecutionEnd endEvent) {}
+  public void end(SparkListenerSQLExecutionEnd endEvent) {
+    log.debug("SparkListenerSQLExecutionEnd - executionId: {}", endEvent.executionId());
+    // TODO: can we get failed event here?
+    // If not, then we probably need to use this only for LogicalPlans that emit no Job events.
+    // Maybe use QueryExecutionListener?
+    endEvent(endEvent.time(), "COMPLETE", null);
+  }
 
   @Override
   public void setActiveJob(ActiveJob activeJob) {}
 
   @Override
   public void start(SparkListenerJobStart jobStart) {
-    log.info("Starting job as part of spark-sql:" + jobStart.jobId());
+    log.debug("SparkListenerJobStart - executionId: " + executionId);
+    jobId = Optional.of(jobStart.jobId());
+    startEvent(jobStart.time());
+  }
+
+  @Override
+  public void end(SparkListenerJobEnd jobEnd) {
+    log.debug("SparkListenerJobEnd - executionId: " + executionId);
+    Exception exception = null;
+    if (jobEnd.jobResult() instanceof JobFailed) {
+      exception = ((JobFailed) jobEnd.jobResult()).exception();
+    }
+    jobId = Optional.of(jobEnd.jobId());
+    endEvent(jobEnd.time(), getEventType(jobEnd.jobResult()), exception);
+  }
+
+  void startEvent(Long time) {
+    if (!started.compareAndSet(false, true)) {
+      log.debug("Start event already emitted: returning");
+      return;
+    }
+
     if (queryExecution == null) {
       log.info("No execution info {}", queryExecution);
       return;
@@ -106,7 +141,7 @@ public class SparkSQLExecutionContext implements ExecutionContext {
     OpenLineage.RunEvent event =
         openLineage
             .newRunEventBuilder()
-            .eventTime(toZonedTime(jobStart.time()))
+            .eventTime(toZonedTime(time))
             .eventType("START")
             .inputs(inputDatasets)
             .outputs(outputDatasets)
@@ -123,7 +158,7 @@ public class SparkSQLExecutionContext implements ExecutionContext {
             .job(buildJob(queryExecution))
             .build();
 
-    log.debug("Posting event for start {}: {}", jobStart, event);
+    log.debug("Posting event for start {}: {}", executionId, event);
     sparkContext.emit(event);
   }
 
@@ -155,9 +190,12 @@ public class SparkSQLExecutionContext implements ExecutionContext {
                     runId, sparkContext.getParentJobName(), sparkContext.getJobNamespace()));
   }
 
-  @Override
-  public void end(SparkListenerJobEnd jobEnd) {
-    log.info("Ending job as part of spark-sql:" + jobEnd.jobId());
+  void endEvent(Long time, String eventType, Exception exception) {
+    if (!finished.compareAndSet(false, true)) {
+      log.debug("Event already finished, returning");
+      return;
+    }
+
     if (queryExecution == null) {
       log.info("No execution info {}", queryExecution);
       return;
@@ -167,6 +205,7 @@ public class SparkSQLExecutionContext implements ExecutionContext {
       log.debug("Traversing optimized plan {}", optimizedPlan.toJSON());
       log.debug("Physical plan executed {}", queryExecution.executedPlan().toJSON());
     }
+
     PartialFunction<LogicalPlan, List<OpenLineage.OutputDataset>> outputVisitor =
         merge(outputDatasetSupplier);
     PartialFunction<LogicalPlan, List<OpenLineage.OutputDataset>> planTraversal =
@@ -175,7 +214,9 @@ public class SparkSQLExecutionContext implements ExecutionContext {
         planTraversal.isDefinedAt(optimizedPlan)
             ? planTraversal.apply(optimizedPlan)
             : Collections.emptyList();
-    outputDatasets = populateOutputMetrics(jobEnd.jobId(), outputDatasets);
+    if (jobId.isPresent()) {
+      outputDatasets = populateOutputMetrics(jobId.get(), outputDatasets);
+    }
 
     List<InputDataset> inputDatasets = getInputDatasets();
     UnknownEntryFacet unknownFacet = unknownEntryFacetListener.build(optimizedPlan).orElse(null);
@@ -183,8 +224,8 @@ public class SparkSQLExecutionContext implements ExecutionContext {
     OpenLineage.RunEvent event =
         openLineage
             .newRunEventBuilder()
-            .eventTime(toZonedTime(jobEnd.time()))
-            .eventType(getEventType(jobEnd.jobResult()))
+            .eventTime(toZonedTime(time))
+            .eventType(eventType)
             .inputs(inputDatasets)
             .outputs(outputDatasets)
             .run(
@@ -194,14 +235,14 @@ public class SparkSQLExecutionContext implements ExecutionContext {
                         new SimpleImmutableEntry(
                             "spark.logicalPlan", buildLogicalPlanFacet(queryExecution.logical())),
                         new SimpleImmutableEntry(
-                            "spark.exception", buildJobErrorFacet(jobEnd.jobResult())),
+                            "spark.exception", buildJobErrorFacet(eventType, exception)),
                         new SimpleImmutableEntry("spark_unknown", unknownFacet),
                         new SimpleImmutableEntry(
                             "spark_version", new SparkVersionFacet(SparkSession.active())))))
             .job(buildJob(queryExecution))
             .build();
 
-    log.debug("Posting event for end {}: {}", jobEnd, event);
+    log.debug("Posting event for end {}: {}", executionId, event);
     sparkContext.emit(event);
   }
 
@@ -276,9 +317,9 @@ public class SparkSQLExecutionContext implements ExecutionContext {
     return LogicalPlanFacet.builder().plan(plan).build();
   }
 
-  protected ErrorFacet buildJobErrorFacet(JobResult jobResult) {
-    if (jobResult instanceof JobFailed && ((JobFailed) jobResult).exception() != null) {
-      return ErrorFacet.builder().exception(((JobFailed) jobResult).exception()).build();
+  protected ErrorFacet buildJobErrorFacet(String jobResult, Exception exception) {
+    if (jobResult.equals("FAIL") && exception != null) {
+      return ErrorFacet.builder().exception(exception).build();
     }
     return null;
   }