+++ integration/spark/src/main/common/java/io/openlineage/spark/agent/lifecycle/plan/InsertIntoDirVisitor.java
@@ -1,13 +1,13 @@
 package io.openlineage.spark.agent.lifecycle.plan;
 
 import io.openlineage.client.OpenLineage;
+import io.openlineage.spark.agent.util.PathUtils;
 import io.openlineage.spark.agent.util.PlanUtils;
 import io.openlineage.spark.agent.util.ScalaConversionUtils;
+import java.net.URI;
 import java.util.Collections;
 import java.util.List;
-import org.apache.hadoop.fs.Path;
-import org.apache.spark.sql.SQLContext;
-import org.apache.spark.sql.catalyst.catalog.CatalogStorageFormat;
+import java.util.Optional;
 import org.apache.spark.sql.catalyst.plans.logical.InsertIntoDir;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
 
@@ -16,26 +16,16 @@ import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
  * OpenLineage.Dataset} being written.
  */
 public class InsertIntoDirVisitor extends QueryPlanVisitor<InsertIntoDir, OpenLineage.Dataset> {
-  private final SQLContext sqlContext;
-
-  public InsertIntoDirVisitor(SQLContext sqlContext) {
-    this.sqlContext = sqlContext;
-  }
 
   @Override
   public List<OpenLineage.Dataset> apply(LogicalPlan x) {
     InsertIntoDir cmd = (InsertIntoDir) x;
-    CatalogStorageFormat storage = cmd.storage();
-    return ScalaConversionUtils.asJavaOptional(storage.locationUri())
+    Optional<URI> optionalUri = ScalaConversionUtils.asJavaOptional(cmd.storage().locationUri());
+    return optionalUri
         .map(
-            uri -> {
-              Path path = new Path(uri);
-              if (uri.getScheme() == null) {
-                path = new Path("file", null, uri.toString());
-              }
-              return Collections.singletonList(
-                  PlanUtils.getDataset(path.toUri(), cmd.child().schema()));
-            })
+            uri ->
+                Collections.singletonList(
+                    PlanUtils.getDataset(PathUtils.fromURI(uri, "file"), cmd.child().schema())))
         .orElse(Collections.emptyList());
   }
 }