+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * dbt: allow parsing artifacts with version higher than officially supported  [@mobuchowski](https://github.com/mobuchowski)
 * dbt: dbt build command is supported  [@mobuchowski](https://github.com/mobuchowski)
 * dbt: fix crash when build command is used with seeds in dbt 1.0.0rc3 [@mobuchowski](https://github.com/mobuchowski)
+* spark: increase logical plan visitor coverage [@mobuchowski](https://github.com/mobuchowski) 
 
 ## [0.3.1](https://github.com/OpenLineage/OpenLineage/releases/tag/0.3.1) - 2021-10-21